modal restructure with children props implementation
