+++ src/pages/surveys/index.tsx
@@ -1,14 +1,8 @@
-import {
-  SurveryModal,
-  SurveyButtonType,
-} from '../../components/modals/surveyModal';
 import { Header } from '../../components/header';
 import { AppLayout } from '../../components/layout/AppLayout';
 import WelcomeMessage from '../../components/welcomeMessage';
-import { useState } from 'react';
 
 function Surveys() {
-  const [publishModalOpen, setPublishModalOpen] = useState<boolean>(true);
   return (
     <AppLayout logo=''>
       <div className='app-container py-2 pl-14 pr-10'>
@@ -20,15 +14,6 @@ function Surveys() {
           />
         </div>
       </div>
-      {publishModalOpen && (
-        <SurveryModal
-          isBackground
-          buttonType={SurveyButtonType.PUBLISH}
-          description='Are you sure you want to publish this survey?'
-          subDescription='This will make the survey available for participants'
-          onClick={() => setPublishModalOpen(false)}
-        />
-      )}
     </AppLayout>
   );
 }
